requirements.txt
flask
numpy
pandas
scikit-learn
gunicorn
joblib
pymongo
bcrypt
python-dotenv